cleanup Plugin.cs

+++ IntroSkipper/Plugin.cs
@@ -112,8 +112,11 @@ public class Plugin : BasePlugin<PluginConfiguration>, IHasWebPages
                     using (var reader = XmlReader.Create(fileStream, settings))
                     {
                         var oldConfig = serializer.Deserialize(reader) as PluginConfiguration;
-                        Instance.UpdateConfiguration(oldConfig);
-                        File.Delete(oldConfigFile);
+                        if (oldConfig != null)
+                        {
+                           Instance.UpdateConfiguration(oldConfig);
+                           File.Delete(oldConfigFile);
+                        }
                     }
                 }
             }
@@ -458,13 +461,13 @@ public class Plugin : BasePlugin<PluginConfiguration>, IHasWebPages
             var config = serverConfiguration.Configuration;
 
             // Get the list of current plugin repositories
-            var pluginRepositories = config.PluginRepositories?.ToList() ?? [];
+            var pluginRepositories = config.PluginRepositories.ToList() ?? [];
 
             // check if old plugins exits
-            if (pluginRepositories.Exists(repo => repo != null && repo.Url != null && oldRepos.Contains(repo.Url)))
+            if (pluginRepositories.Exists(repo => repo.Url != null && oldRepos.Contains(repo.Url)))
             {
                 // remove all old plugins
-                pluginRepositories.RemoveAll(repo => repo != null && repo.Url != null && oldRepos.Contains(repo.Url));
+                pluginRepositories.RemoveAll(repo => repo.Url != null && oldRepos.Contains(repo.Url));
 
                 // Add repository only if it does not exit
                 if (!pluginRepositories.Exists(repo => repo.Url == "https://manifest.intro-skipper.org/manifest.json"))
@@ -500,8 +503,7 @@ public class Plugin : BasePlugin<PluginConfiguration>, IHasWebPages
         string searchPattern = "dashboard-dashboard.*.chunk.js";
         string[] filePaths = Directory.GetFiles(webPath, searchPattern, SearchOption.TopDirectoryOnly);
         string pattern = @"buildVersion""\)\.innerText=""(?<buildVersion>\d+\.\d+\.\d+)"",.*?webVersion""\)\.innerText=""(?<webVersion>\d+\.\d+\.\d+)";
-        string buildVersionString = "unknow";
-        string webVersionString = "unknow";
+        string webVersionString = "unknown";
         // Create a Regex object
         Regex regex = new Regex(pattern);
 
@@ -514,14 +516,13 @@ public class Plugin : BasePlugin<PluginConfiguration>, IHasWebPages
             // search for buildVersion and webVersion
             if (match.Success)
             {
-                buildVersionString = match.Groups["buildVersion"].Value;
                 webVersionString = match.Groups["webVersion"].Value;
                 _logger.LogInformation("Found jellyfin-web <{WebVersion}>", webVersionString);
                 break;
             }
         }
 
-        if (webVersionString != "unknow")
+        if (webVersionString != "unknown")
         {
             // append Revision
             webVersionString += ".0";
